requirements.txt
awscli==1.25.86
botocore==1.27.85
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1
django-ckeditor==6.5.0
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-csvimport==2.16
django-embed-video==1.4.4
django-form-utils==1.0.3
django-import-export==2.8.0
django-js-asset==2.0.0
docutils==0.16
et-xmlfile==1.1.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
platformdirs==2.5.2
pyasn1==0.4.8
python-dateutil==2.8.2
PyYAML==6.0
reportlab==3.6.11
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tk==0.1.0
urllib3==1.26.12
virtualenv==20.16.5
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
